OS X build fixes
Changes in Rust 0.11-pre may have necessitated the attribute
modifications in common/lib.rs

+++ arch/x86/drivers/pic.rs
@@ -16,7 +16,7 @@ pub fn remap() {
 
 pub fn enable(irq: u8) {
     let port: u16 = if (irq & 0b1000) == 0 { 0x21 } else { 0xa1 };
-    let mask: u8 = !(1u8 << (irq & 0b111));
+    let mask: u8 = !(1u8 << (irq & 0b111) as uint);
 
     io::out(port, io::inb(port) & mask);
 }
